Factor macro_rules! highlighting out
